Address code review findings for SIMD optimizations
- Upgrade StructuralIndex to u64 for >4GB document support
- Add UnclosedComment error for block comments not closed at EOF
- Document scratch buffer lifetime and event pinning semantics
- Add AVX2 path tests for 32-byte boundaries and 64-byte blocks
- Extract SIMD_BLOCK_SIZE constant (64 bytes)
- Document ScanState invariants and valid state combinations
- Document MAX_TRACKED_LINES overflow behavior in constants.zig
- Add IndexParser-specific fuzz target comparing strategies
- Export ParseStrategy, parseReader, parseReaderWithOptions from root

+++ src/simd/structural.zig
@@ -170,7 +170,7 @@ test "StructuralIndex basic scan" {
     }
 }
 
-fn containsIndex(indices: []const u32, pos: usize) bool {
+fn containsIndex(indices: []const u64, pos: usize) bool {
     for (indices) |idx| {
         if (idx == pos) return true;
     }
@@ -239,5 +239,5 @@ test "StructuralIndex scanReader matches scan" {
     defer allocator.free(rebuilt_slice);
 
     try std.testing.expectEqualStrings(source, rebuilt_slice);
-    try std.testing.expectEqualSlices(u32, direct.slice(), streamed.index.slice());
+    try std.testing.expectEqualSlices(u64, direct.slice(), streamed.index.slice());
 }
